requirements.txt
praw==7.7.1
prawcore==2.3.0
python-dotenv==1.0.0
